test: split native provider adapter tests by provider (#363)
